+++ app/(dashboard)/page.tsx
@@ -1,0 +1,12 @@
+import { HomeTemplate } from '@/components';
+
+export default function PageRoute() {
+  return (
+    <>
+      <p className='text-4xl text-blue-600'>MAIN</p>
+      <HomeTemplate />
+    </>
+  );
+}
+
+// <main className='border-t border-brand-200 bg-brand-100 dark:border-brand-600 dark:bg-brand-800'>
