--- app/lib/get-session.ts
@@ -1,6 +1,0 @@
-import { authOptions } from '@/pages/api/auth/[...nextauth]';
-import { getServerSession } from 'next-auth';
-
-export async function getSession() {
-  return await getServerSession(authOptions);
-}
