--- app/lib/get-user.ts
@@ -1,25 +1,0 @@
-import { getSession } from './get-session';
-import db from './prisma';
-
-export async function getUser() {
-  try {
-    const session = await getSession();
-    if (!session?.user?.email) return null;
-
-    const user = await db.user.findUnique({
-      where: {
-        email: session.user.email,
-      },
-    });
-    if (!user) return null;
-
-    return {
-      ...user,
-      createdAt: user.createdAt.toISOString(),
-      updatedAt: user.updatedAt.toISOString(),
-      emailVerified: user.emailVerified?.toISOString() || null,
-    };
-  } catch (error) {
-    return null;
-  }
-}
